Update Alias i2c-b.dts
Make alias i2c0 to match the RPi pin 27/28.

This also relates to the use of aliases in device tree, the lowest defined alias gets "i2c-0", no mater what number the lowest alias number is.  to use RPi intended code, this overlay must be loaded as well to get the pin 3 and 5 ic2 to be i2c-1.